backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0
aiofiles>=23.2.0
openai>=1.12.0
ruff>=0.2.0
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
